fixed some details

+++ fun2travel/Models/ViewModels/AdventuresVM.cs
@@ -9,8 +9,8 @@ namespace fun2travel.Models.ViewModels
     {
         public int Id { get; set; }
         public string ActivityName { get; set; }
-        public int ActivityPrice { get; set; }
-        public int ActivityRentalPrice { get; set; }
+        public decimal ActivityPrice { get; set; }
+        public decimal ActivityRentalPrice { get; set; }
         public bool EquipmentCanBeRented { get; set; }
         public string ActivityDescription { get; set; }
         public string ActivityPic1 { get; set; }
